always use dark theme
It causes multiple rerenders when using the st-theme package which impact performance. Currently there is no proper way to read out the current theme when it is not explicitly set. Streamlit team seems to work on a feature which might fix this, but it is delayed. So we currently always do dark theme for now.